// Backend-Service/src/src-gate/libs/constants/enums.ts
export enum UserRoles {
    MANAGER = 'Manager',
    TEAM_LEADER = 'Team Leader',
    EMPLOYEE = 'Employee',
}

export enum TaskStatuses {
    COMPLETED = 'Completed',
    IN_PROGRESS = 'In Progress',
    PENDING = 'Pending'
}

export enum ErrTypes {
    CUSTOM = "Custom"
}

export enum errMessages {
    INVALID_ROLE = 'You do not have permission for this operation!'
}

export enum ClientDomainUrls {
    LOCALHOST = "http://localhost:4200"
}